+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this theme will be documented in this file.
   - Update `statusbar` color for `debugging` state
   - Update `statusbar` color for `nofolder` state
   - Bring back markdown syntax highlighting
+  - Highlight `jsdoc` keyword
 
   ### Changed
   - `normal` font-style is not supported. Replace it with an empty string.